pyproject.toml
[tool.poetry]
name = "students-api"
version = "0.1.0"
description = ""
authors = ["Ahmed Hossam"]
readme = "README.md"
packages = [{include = "students_api"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.2"
djangorestframework = "^3.14.0"
django-rest-framework = "^0.1.0"
drf-yasg = "^1.21.4"
jwt = "^1.3.1"
pyjwt = "^2.6.0"
djangorestframework-simplejwt = "^5.2.2"
pyyaml = "^6.0"
python-dotenv = "^0.21.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
